extract utils for resave_form, resave_case and move resync_case_to_es
resync_case_to_es should belong to case module and not reports, closest i could find was hqcase
