+++ corehq/apps/hqadmin/management/commands/check_case_index_ids.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import csv
 from django.core.management import BaseCommand, call_command
 from corehq.apps.receiverwrapper.util import get_app_version_info
-from corehq.apps.reports.util import resync_case_to_es
+from corehq.apps.hqcase.utils import resync_case_to_es
 from corehq.apps.users.util import cached_owner_id_to_display
 from corehq.elastic import ES_MAX_CLAUSE_COUNT
 from corehq.apps.es.cases import CaseES
